+++ CHANGELOG.md
@@ -5,6 +5,12 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- `circulant` and `companion` can build complex matrices.
+
 ## [0.1.1] - 2020-06-23
 
 ### Added
@@ -21,5 +27,6 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - A circulent matrix builder.
 - A companion matrix builder.
 
+[Unreleased]: https://github.com/vinesystems/lair/compare/0.1.1...master
 [0.1.1]: https://github.com/vinesystems/lair/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/vinesystems/lair/tree/0.1.0